Fix property removal

+++ plugin/src/main/java/io/github/zrdzn/minecraft/greatlifesteal/config/bean/beans/ActionBean.java
@@ -8,7 +8,7 @@ public class ActionBean {
 
     private boolean enabled = false;
     private ActionType type = ActionType.DISPATCH_COMMANDS;
-    private double activateAtHealth = 4;
+    private double activateAtHealth = 4.0D;
     private List<String> parameters = Collections.singletonList("gamemode spectator {victim}");
 
     public boolean isEnabled() {
@@ -31,7 +31,7 @@ public class ActionBean {
         return this.activateAtHealth;
     }
 
-    public void setActivateAtHealth(int activateAtHealth) {
+    public void setActivateAtHealth(double activateAtHealth) {
         this.activateAtHealth = activateAtHealth;
     }
